First pass implementation of generating CPLs conforming to the 2013 and 2016 schemas and the first implementation of an IMP builder to generate AssetMap, PKL and CPL documents. Currently the assumptions we make about the IMP are that it will be a single segment and a complete IMP or at the very least the header partition corresponding to all the IMF Track files that are a part of all the Virtual Tracks of the composition are available.

+++ src/test/java/com/netflix/imflibrary/writerTools/AssetMapBuilderFunctionalTest.java
@@ -21,10 +21,12 @@ import com.netflix.imflibrary.IMFErrorLogger;
 import com.netflix.imflibrary.IMFErrorLoggerImpl;
 import com.netflix.imflibrary.RESTfulInterfaces.IMPValidator;
 import com.netflix.imflibrary.RESTfulInterfaces.PayloadRecord;
+import com.netflix.imflibrary.exceptions.IMFAuthoringException;
 import com.netflix.imflibrary.st0429_9.AssetMap;
 import com.netflix.imflibrary.utils.ErrorLogger;
 import com.netflix.imflibrary.utils.FileByteRangeProvider;
 import com.netflix.imflibrary.utils.ResourceByteRangeProvider;
+import com.netflix.imflibrary.utils.Utilities;
 import com.netflix.imflibrary.writerTools.utils.IMFUtils;
 import org.testng.Assert;
 import org.testng.annotations.Test;
@@ -48,6 +50,7 @@ public class AssetMapBuilderFunctionalTest {
 
     @Test
     public void assetMapBuilderTest() throws IOException, SAXException, JAXBException, URISyntaxException {
+
         File inputFile = TestHelper.findResourceByPath("TestIMP/NYCbCrLT_3840x2160x23.98x10min/ASSETMAP.xml");
         ResourceByteRangeProvider resourceByteRangeProvider = new FileByteRangeProvider(inputFile);
         IMFErrorLogger imfErrorLogger = new IMFErrorLoggerImpl();
@@ -84,11 +87,25 @@ public class AssetMapBuilderFunctionalTest {
         }
 
         IMFErrorLogger assetMapBuilderErrorLogger = new IMFErrorLoggerImpl();
-        File assetMapFile = new AssetMapBuilder(assetMap.getUUID(), annotationText, creator, issueDate, issuer, assetMapBuilderAssets, tempDir, assetMapBuilderErrorLogger).build();
+        List<ErrorLogger.ErrorObject> errors = new AssetMapBuilder(assetMap.getUUID(), annotationText, creator, issueDate, issuer, assetMapBuilderAssets, tempDir, assetMapBuilderErrorLogger).build();
+
+        if(imfErrorLogger.getErrors(IMFErrorLogger.IMFErrors.ErrorLevels.FATAL, 0, imfErrorLogger.getNumberOfErrors()).size() > 0){
+            throw new IMFAuthoringException(String.format("Fatal errors occurred while generating the AssetMap. Please see following error messages %s", Utilities.serializeObjectCollectionToString(imfErrorLogger.getErrors(IMFErrorLogger.IMFErrors.ErrorLevels.FATAL, 0, imfErrorLogger.getNumberOfErrors()))));
+        }
+
+        File assetMapFile = null;
+        for(File file : tempDir.listFiles()){
+            if(file.getName().contains("AssetMap-")){
+                assetMapFile = file;
+            }
+        }
+        if(assetMapFile == null){
+            throw new IMFAuthoringException(String.format("AssetMap file does not exist in the working directory %s, IMP is incomplete", tempDir.getAbsolutePath()));
+        }
         Assert.assertTrue(assetMapFile.length() > 0);
 
-        List<ErrorLogger.ErrorObject> errors = IMPValidator.validateAssetMap(new PayloadRecord(new FileByteRangeProvider(assetMapFile).getByteRangeAsBytes(0, assetMapFile.length()-1), PayloadRecord.PayloadAssetType.AssetMap, 0L, 0L));
-        Assert.assertTrue(errors.size() == 0);
+        List<ErrorLogger.ErrorObject> assetMapValidationErrors = IMPValidator.validateAssetMap(new PayloadRecord(new FileByteRangeProvider(assetMapFile).getByteRangeAsBytes(0, assetMapFile.length()-1), PayloadRecord.PayloadAssetType.AssetMap, 0L, 0L));
+        Assert.assertTrue(assetMapValidationErrors.size() == 0);
 
         //Destroy the temporary working directory
         tempDir.delete();
